requirements.txt
kafka-python
mysqlclient
six